RtAudio => OpenAL

+++ build_vs_debug.bat
@@ -1,6 +1,9 @@
 @echo off
 rem Delete the build directory if it exists
-rmdir /S /Q build
+if exist build rmdir /S /Q build
+
+rem Run vcpkg install with the correct triplet
+vcpkg install --triplet x64-windows
 
 rem Run CMake to configure the project and generate build files
 cmake -S . -B build -G "Visual Studio 17 2022" -DCMAKE_TOOLCHAIN_FILE="C:/vcpkg/scripts/buildsystems/vcpkg.cmake"
